backend/pyproject.toml
[tool.poetry]
name = "api"
version = "0.1.0"
description = ""
authors = ["krystasis"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
django = "^4.2"
django-rest-framework = "^0.1.0"
django-filter = "^23.2"
environ = "^1.0"
django-cors-headers = "^3.14.0"
django-environ = "^0.10.0"
djangorestframework-jwt = "^1.11.0"
requests = "^2.29.0"
djangorestframework-simplejwt = "^5.2.2"


[tool.poetry.group.dev.dependencies]
black = "^23.3.0"
isort = "^5.12.0"
pydocstringformatter = "^0.7.3"
pytest = "^7.3.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
